+++ src/commands/__init__.py
@@ -1,5 +1,5 @@
 from typing import Callable
-from . import help, generate, update, path
+from . import help, generate, update, path, clean
 from .generate import NAME as DEFAULT_COMMAND_NAME
 
 COMMANDS: dict[str, Callable[[str, list[str]], None]] = {
@@ -7,4 +7,5 @@ COMMANDS: dict[str, Callable[[str, list[str]], None]] = {
     help.NAME: help.execute,
     update.NAME: update.execute,
     path.NAME: path.execute,
+    clean.NAME: clean.execute,
 }