+++ src/web/__init__.py
@@ -1,3 +1,4 @@
 """Contains miscellaneous utility functions used by multiple other modules."""
 from .cache_entry import Expiration
+from .cache import clear_cache
 from .functions import get, download